Use OneOf to handle responses better (#24)
* Replaces exception based approach in responses for a (kind of) discriminated union approach.
Clients should not be affected by this change.

* Updates package-lock.json.

+++ LetsTalk.Application/Roles/Commands/DeleteRoleCommand.cs
@@ -1,11 +1,13 @@
 ﻿using FluentValidation;
-using LetsTalk.Exceptions;
+using LetsTalk.Errors;
 using LetsTalk.Repositories;
+using LetsTalk.Services;
 using MediatR;
+using OneOf.Types;
 
 namespace LetsTalk.Roles.Commands;
 
-public sealed class DeleteRoleCommand : IRequest
+public sealed class DeleteRoleCommand : IRequest<Response<Success>>
 {
     public required string Id { get; init; }
 
@@ -17,19 +19,30 @@ public sealed class DeleteRoleCommand : IRequest
         }
     }
 
-    public sealed class Handler : IRequestHandler<DeleteRoleCommand>
+    public sealed class Handler : IRequestHandler<DeleteRoleCommand, Response<Success>>
     {
+        private readonly IValidatorService<DeleteRoleCommand> _validator;
         private readonly IRoleRepository _roleRepository;
 
-        public Handler(IRoleRepository roleRepository)
+        public Handler(IValidatorService<DeleteRoleCommand> validator, IRoleRepository roleRepository)
         {
+            _validator = validator;
             _roleRepository = roleRepository;
         }
 
-        public async Task Handle(DeleteRoleCommand request, CancellationToken cancellationToken)
+        public async Task<Response<Success>> Handle(DeleteRoleCommand request, CancellationToken cancellationToken)
         {
-            var role = await _roleRepository.GetByIdAsync(request.Id, cancellationToken) ?? throw ExceptionFor<Role>.NotFound(r => r.Id, request.Id);
+            var validation = await _validator.ValidateAsync(request, cancellationToken);
+            if (!validation.IsValid)
+                return new Invalid(validation.ToDictionary());
+
+            var role = await _roleRepository.GetByIdAsync(request.Id, cancellationToken);
+            if (role is null)
+                return new NotFound();
+
             await _roleRepository.DeleteAsync(role, cancellationToken);
+
+            return new Success();
         }
     }
 }
